backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
pydantic
pydantic-settings
openai
python-dotenv

psycopg2-binary

